requirements.txt
Django==1.8.5
argparse==1.2.1
distribute==0.6.24
django-extensions==1.5.9
django-mptt==0.7.4
django-tastypie==0.12.2
psycopg2==2.6.1
pyparsing==2.0.6
python-dateutil==2.4.2
python-mimeparse==0.1.4
pyzmq==14.7.0
six==1.10.0
wsgiref==0.1.2
